fix[graphql]: test

+++ backend/molgenis-emx2-graphql/src/test/java/org/molgenis/emx2/graphql/TestGraphqlDatabaseFields.java
@@ -33,7 +33,9 @@ public class TestGraphqlDatabaseFields {
   @Test
   public void testCreateAndDeleteSchema() throws IOException {
     // ensure schema doesn't exist
-    database.dropSchema(schemaName + "B");
+    if (database.getSchema(schemaName + "B") != null) {
+      database.dropSchema(schemaName + "B");
+    }
 
     int realLength = database.getSchemaNames().size();
     int length = execute("{Schemas{name}}").at("/data/Schemas").size();
